+++ libs/cardano-data/testlib/Test/Cardano/Data/Arbitrary.hs
@@ -4,13 +4,17 @@
 
 module Test.Cardano.Data.Arbitrary where
 
+import Data.Map.Strict qualified as Map
 import Data.OMap.Strict qualified as OMap
 import Data.OSet.Strict qualified as OSet
+import Data.Set qualified as Set
+import Lens.Micro (set)
 import Test.Cardano.Ledger.Binary.Arbitrary ()
 import Test.QuickCheck
 
 instance (Arbitrary a, Ord a) => Arbitrary (OSet.OSet a) where
-  arbitrary = OSet.fromSet <$> arbitrary
+  arbitrary = fmap (OSet.fromSet . Set.fromList) . shuffle . Set.toList =<< arbitrary
 
-instance (Ord k, Ord v, Arbitrary v, OMap.HasOKey k v) => Arbitrary (OMap.OMap k v) where
-  arbitrary = OMap.fromSet <$> arbitrary
+instance (Ord v, Arbitrary v, OMap.HasOKey k v, Arbitrary k) => Arbitrary (OMap.OMap k v) where
+  arbitrary =
+    fmap OMap.fromFoldable . shuffle . Map.elems . Map.mapWithKey (flip (set OMap.okeyL)) =<< arbitrary